Create and solve army page
